Revert Add `File.equal` check to `LocalBackup`
Reverts commit 8d53d1415850f0c538bfee1a7c086bcbfce1c4eb but keeping the
refactoring.

Hmm this doesn't add much for `LocalBackup`. Just makes it (about 3x)
slower. We're doing the backup frequently so if there are corrupted
files (if that is even a possible outcome of `Files.copy`) they'll be
updated the next day.

Makes more sense for `S3Backup` where it's not run as often, there's less
(and smaller) files, and the transmission is over the internet where a
checksum is more valuable. Not to mention AWS runs it on their end...