Revert "Observe performance test failure by running backup sequentially"
This reverts commit d705efbe9b5f9586539f3e7c7a40ba6d3ae8c84b.

Still passes on the agents... but still observed a slow down.

Locally there is a significant slowdown (running on 16c32t 5950x).

Because the agents are only 2 core, the difference between seq/parallel
is not that much (though it's mainly IO ops???). Previously had to raise
the threshold to let it pass on the agents (df279fbbfe64d0d0a0b3bc7e99db5ef18078403f),
because they're pretty much seq. Also had to remove the 10,000 file test
where parallel has more impact.

Basically... performance test running on github won't pick up someone
messing with the parallelism. But it is needed because in practice
100,000 files are being backed up.