Handle copying to non-empty dir in `LocalBackup`
Could happen if on the first run a file `A/B/C/D` was copied.
Then on the next run the file was `A/B/C`; the non-empty dir `C/D` needs
to be overwritten.

Could happen on `S3Backup` also? Not sure how AWS deals with it...

It's rare a file would match a dir name (i.e. no extension)...