Skip put of empty dirs to S3
Turns out the folders in S3 are deleted when all the objects within the
folder is deleted; but only if the folders were created with the
objects, i.e. not created separately.

Therefore we don't have to handle the cleanup of empty folders if we
don't create them.

However - I also discovered S3 ListObjects does return empty folders
(essentially it returns all leaves) so we may as well implement it?

We need to change `Path` back to `String` since S3 identifies folders
via trailing `/`? Or just some way to identify we're deleting a folder.