Try sync attributes after delete/put
The problem is that parent dir last-modified changes when files are
copied in, so on subsequent runs the dir is incorrectly deleted (fails
`same` check).

The idea here was to sync the last-modified time at the end of the run,
however it isn't working??? It works when just doing the sync in
isolation, but not on a complete run??? Is there some lag to sending the
request over the network so they get out of order or something??

New idea is to change the algo so doesn't delete dirs with mismatch
last-modified time.