Remove `BackupRunner`; put algorithm in abstract `Backup` class instead
The abstract class is a cleaner design tbh... the `public` methods on
the interface were exposing too much. Also `Backup.run` seems natural.

`BackupRunner` only existed for the unit test, but that test has been
quite fragile. Plus it's already covered by the integration test, which
is a better fit for this code anyway.

(Usually `abstract class` is a smell... this seems like one of those
rare cases where it's right? https://stackoverflow.com/a/2947823/6122976)